Changed the definition of `execution`
The previous definition of `execution` was not correct. It used to say "process of parsing code in an interpreter-defined way", when it is not at all *parsing* of code (parsing, according to this standard, is an entirely different thing). Plus it is *not* the job of the interpreter to document how a loaded code is executed either